Add header for base input manager
